+++ lib/playx_theme.dart
@@ -1,11 +1,13 @@
 library playx_theme;
 
 export 'src/playx_theme.dart';
-
-///Library exports
 export 'src/builder/builder.dart';
 export 'src/config/config.dart';
 export 'src/model/x_color_scheme.dart';
 export 'src/model/x_theme.dart';
-export 'package:playx_core/playx_core.dart';
+export 'src/widgets/image_theme.dart';
+export 'src/utils/utils.dart';
 
+
+export 'package:playx_core/playx_core.dart';
+export 'package:flex_seed_scheme/flex_seed_scheme.dart';